features work together properly
