--- src/com/codecool/foodswap/model/Rank.java
@@ -1,10 +1,0 @@
-package com.codecool.foodswap.model;
-
-public enum Rank {
-
-    KITCHEN_HELPER,
-    CHEF,
-    MASTER_CHEF,
-    GORDON_RAMSAY,
-
-}
