+++ src/main/java/com/codecool/foodswap/model/Rank.java
@@ -1,4 +1,10 @@
 package com.codecool.foodswap.model;
 
 public enum Rank {
+
+    KITCHEN_HELPER,
+    CHEF,
+    MASTER_CHEF,
+    GORDON_RAMSAY,
+
 }
